Splitting out dnsutil into separate dig module
